backend_hack2/backend/requirements.txt
# Core dependencies
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx>=0.28.1
motor>=3.6.0
authlib>=1.3.0
email-validator>=2.1.0

# AI and ML
groq>=0.9.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
google-genai>=1.52.0

# Audio/Video (if needed)
elevenlabs>=1.0.0
sounddevice>=0.5.0
pydub>=0.25.0

# Utilities
python-dotenv>=1.0.1
streamlit>=1.40.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
mongomock>=4.1.0
black>=24.0.0
mypy>=1.10.0
pre-commit>=3.7.0
slowapi>=0.1.9
redis>=5.0.0
